+++ pkgs/native_assets_cli/test/api/build_test.dart
@@ -34,7 +34,6 @@ void main() async {
       ..setupHookConfig(
         packageRoot: tempUri,
         packageName: packageName,
-        targetOS: OS.iOS,
         buildAssetTypes: ['foo'],
       )
       ..setupBuildConfig(