+++ pkgs/native_toolchain_c/test/clinker/objects_test.dart
@@ -36,12 +36,12 @@ Future<void> main() async {
         buildAssetTypes: [CodeAsset.type],
         packageName: 'testpackage',
         packageRoot: tempUri,
-        targetOS: os,
       )
       ..setupLinkConfig(
         assets: [],
       )
       ..setupCodeConfig(
+        targetOS: os,
         targetArchitecture: architecture,
         linkModePreference: LinkModePreference.dynamic,
         cCompilerConfig: cCompiler,